+++ docs/src/examples/components/Accordion/Types/AccordionExample.shorthand.tsx
@@ -3,10 +3,12 @@ import { Accordion } from '@stardust-ui/react'
 
 const panels = [
   {
+    key: 'p',
     title: 'What is a point?',
     content: 'Use the haptic SDD circuit, then you can index the redundant pixel!',
   },
   {
+    key: 'd',
     title: 'What is a dimension of a point?',
     content: 'We need to copy the primary USB firewall!',
   },