+++ docs/src/examples/components/Accordion/Types/AccordionExclusiveExample.shorthand.tsx
@@ -3,14 +3,17 @@ import { Accordion } from '@stardust-ui/react'
 
 const panels = [
   {
+    key: 'one',
     title: 'One',
     content: '2 3 4',
   },
   {
+    key: 'two',
     title: 'Five',
     content: '6 7 8 9',
   },
   {
+    key: 'three',
     title: "What's next?",
     content: '10',
   },